+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 * `Settings.IsEqualTo()` method will compare objects correctly [#291](https://github.com/NetOfficeFw/NetOffice/issues/291)
 * `Settings` implements `IEquatable<Settings>` in favor of old `IsEqualTo()` method
+* `ResourceUtils.ReadImage()` will correctly read the image from resources [#292](https://github.com/NetOfficeFw/NetOffice/issues/292)
 
 ### Bug Fixes
 * Fix [#291](https://github.com/NetOfficeFw/NetOffice/issues/291) _`Settings.IsEqualTo` compares `EnableSafeMode` incorrectly_
+* Fix [#292](https://github.com/NetOfficeFw/NetOffice/issues/292) _`ResourceUtils.ReadImage()` calls itself recursively_
 
 ## v1.7.7